update-20231107

+++ app/Http/Controllers/Api/TypeController.php
@@ -16,7 +16,19 @@ class TypeController extends Controller
     public function index()
     {
         // return TypeListResource::collection(Type::query()->paginate(10));
-        return TypeListResource::collection(Type::with('category')->paginate(10));
+        $search = request('search', false);
+        $perPage = request('per_page', 10);
+        $sortField = request('sort_field', 'updated_at');
+        $sortDirection = request('sort_direction', 'desc');
+
+        $query = Type::query();
+        $query->orderBy($sortField, $sortDirection);
+        if($search) {
+            $query->where('code', 'like', "%{$search}%")
+            ->orWhere('name', 'like', "%{$search}%");
+        }
+
+        return TypeListResource::collection($query->paginate($perPage));
     }
 
     /**
@@ -24,7 +36,13 @@ class TypeController extends Controller
      */
     public function store(TypeRequest $request)
     {
-        return new TypeResource(Type::create($request->validated()));
+        $data = $request->validated();
+        $data['created_by'] = $request->user()->id;
+        $data['updated_by'] = $request->user()->id;
+
+        $type = Type::create($data);
+
+        return new TypeResource($type);
     }
 
     /**
@@ -41,6 +59,9 @@ class TypeController extends Controller
     public function update(TypeRequest $request, Type $type)
     {
         $type->update($request->validated());
+        $data['updated_by'] = $request->user()->id;
+
+        $type->update($data);
         return new TypeResource($type);
     }
 
@@ -49,6 +70,13 @@ class TypeController extends Controller
      */
     public function destroy(Type $type)
     {
+        // Cek jika Unit memiliki Product
+        if ($unit->product()->count() > 0) {
+            // Lempar Exception dengan pesan error
+            throw new \Exception("Unit ini memiliki tipe terkait dan tidak dapat dihapus.");
+            // return response()->json(['error' => 'Kategori ini memiliki tipe terkait dan tidak dapat dihapus.'], 409);
+        }
+        
         $type->delete();
         return response()->noContent();
     }
